+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/Orange-OpenSource/ouds-ios-design-system-toolbox/compare/1.1.0...develop)
 
+### Changed
+
+- Rename of *Orange Business Tools*  theme to *Orange Compact* (Orange-OpenSource/ouds-ios#1292)
+
+
 ## [1.1.0](https://github.com/Orange-OpenSource/ouds-ios-design-system-toolbox/compare/1.0.0...1.1.0) - 2026-01-27
 
 ### Added